chore(webview): extract vscode accessor into services/vscode-api
Moves the `window.acquireVsCodeApi?.()` initialisation out of `main.tsx`
into a side-effect-light module. `main.tsx` still re-exports `vscode` so
existing imports (`from '../main'`) keep working unchanged, but the new
`ccwf preview` entry can depend on the channel without auto-mounting the
full canvas via main.tsx's createRoot.
